trading/requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.7
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
djangorestframework-simplejwt==4.8.0
flake8==4.0.1
Flask==2.0.2
flower==0.9.7
humanize==3.12.0
idna==3.3
importlib-metadata==4.2.0
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==4.6.11
MarkupSafe==2.0.1
mccabe==0.6.1
MyApplication==0.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.22
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
typing==3.7.4.3
typing_extensions==4.0.0
urllib3==1.26.7
vine==1.3.0
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.6.0
psycopg2-binary==2.8.5
